fix(rooms): record signed-in room history per presented question, not at start [STE-273]
Codex review round 3 on PR #161.

Allocating the entire preselected pool as "seen" for every signed-in
player at room start marked questions a player might never reach (an
abandoned game or an early leaver), wrongly deprioritizing them and
advancing their cooldown cycle. This mirrors the flaw the guest and solo
paths already avoid by recording only presented questions.

- Remove the start-time seen_questions bulk insert.
- Generalize the room recording hook (useRecordGuestRoomQuestion ->
  useRecordRoomQuestion): guests still write browser-local history;
  signed-in players now POST /api/questions/seen per presented question,
  so only what was actually shown is recorded. The endpoint's 24h replay
  guard keeps retries/rejoins from inflating counts.

Room selection still unions every participant's server-side history at
start; only the write path changed.
